+++ documentation/ReleaseNotes.md
@@ -13,6 +13,7 @@ Add:
 
 Fix:
 
+- Hiding Home Page tabs no longer cuts the top thumbnails;
 - Selector being used to hide the Sleep Timer element is properly defined now;
 
 ## 1.11.0